clean and sort functions
